requirements.txt
arrow==1.1.0
asgiref==3.3.4
binaryornot==0.4.4
bpmappers==1.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cond==1.2.5
cookiecutter==1.7.2
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20121119
Django==3.2
django-allauth==0.44.0
django-contrib-comments==2.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.7.0
django-crontab==0.7.1
django-formtools==2.1
django-import-export==0.6.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-redis==5.0.0
django-reversion==2.0.12
django-simple-captcha==0.5.14
django-taggit==1.4.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
future==0.18.2
id-validator==1.0.20
idna==2.10
jdcal==1.3
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
mysqlclient==2.0.1
numpy==1.20.3
oauthlib==3.1.0
odfpy==1.3.6
opencv-python==4.5.2.52
openpyxl==2.4.9
optionaldict==0.1.2
pandas==1.2.4
Pillow==8.0.1
poyo==0.5.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2018.4
PyYAML==3.12
redis==3.5.3
reportlab==3.5.56
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.6.3
six==1.11.0
sqlparse==0.4.1
svgwrite==1.4.1
tablib==0.12.1
text-unidecode==1.3
threadpoolctl==2.1.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
Tree==0.2.4
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.4
wechatpy==1.8.15
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.12.0
zmail==0.2.8
